Provide better error reporting
When we have the power of Q there isn't much need to continue the
callback style.

Also the former method would cause issues if the return body from github
was bad JSON data (throwing a parse error). Since the logic is now in a
promise function it can be properly managed via Q.